QuickSaveButton 컴포넌트 추가 및 스레드 항목에 통합
